+++ EngineTester/SponzaDemo.cpp
@@ -47,12 +47,12 @@ void SponzaDemo::Init(Scene *scene)
     lightMat.mKd = vec3(1.0f);
     lightMat.mNeedsForwardRender = true;
     scene->mMaterials.push_back(lightMat);
-    const int lightMatIdx = scene->mMaterials.size() - 1;
+    const int lightMatIdx = static_cast<int>(scene->mMaterials.size()) - 1;
 
     TriangleMesh lightMesh = GetSphereMeshSolid(false, 1, 5.0f);
     lightMesh.FlipNormals();
     scene->mTriMeshes.push_back(lightMesh);
-    const int lightMeshIdx = scene->mTriMeshes.size() - 1;
+    const int lightMeshIdx = static_cast<int>(scene->mTriMeshes.size()) - 1;
 
     scene->AddObject(AssetLoader::Get().CreateObject(scene, lightMeshIdx, lightMatIdx));
     mPLightObj = &scene->GetObject(scene->GetNumObjects() - 1);
